[server] Split main package into pkg/server
