+++ doc/cqlsh.md
@@ -2,3 +2,7 @@
 
 cqlsh defined commands
 
+The DESCRIBE or DESC command outputs information about the connected Cassandra cluster, or about the data stored on it. 
+To query the system tables directly, use SELECT
+
+https://docs.datastax.com/en/cql/3.1/cql/cql_reference/describe_r.html
