+++ doc/schema.md
@@ -1,0 +1,5 @@
+# Schema
+
+How to query schema using CQL (not cqlsh)
+
+- gocql has `KeyspaceMetadata` see https://github.com/gocql/gocql/blob/master/metadata.go
