+++ pkg/server/pkg.go
@@ -1,0 +1,7 @@
+package server
+
+import (
+	dlog "github.com/dyweb/gommon/log"
+)
+
+var log = dlog.NewLibraryLogger()
