+++ ui/Makefile
@@ -1,0 +1,3 @@
+.PHONY: serve
+serve:
+	npm run serve
